sb-requirements.txt
django_summernote
django-pinax-referral
django-crontab
